@@ -1,14 +1,12 @@
 // Copyright (c) .NET Foundation. All rights reserved. Licensed under the Apache License, Version
 // 2.0. See License.txt in the project root for license information.
 
-using System;
-using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Microsoft.AspNetCore.SignalR.Protocol;
-using Orleans;
 using SignalR.Orleans.Tests.Models;
 using SignalR.Orleans.Tests.AspnetSignalR;
 using Xunit;
+using System.Diagnostics;
 
 namespace SignalR.Orleans.Tests
 {
@@ -24,8 +22,8 @@ namespace SignalR.Orleans.Tests
         [Fact]
         public async Task AddGroupAsync_ForConnection_OnDifferentServer_AlreadyInGroup_SkipsDuplicate()
         {
-            var manager1 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
-            var manager2 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
+            var manager1 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
+            var manager2 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
 
             using (var client = new TestClient())
             {
@@ -46,8 +44,8 @@ namespace SignalR.Orleans.Tests
         [Fact]
         public async Task AddGroupAsync_ForConnection_OnDifferentServer_Works()
         {
-            var manager1 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
-            var manager2 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
+            var manager1 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
+            var manager2 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
 
             using (var client = new TestClient())
             {
@@ -66,7 +64,7 @@ namespace SignalR.Orleans.Tests
         [Fact]
         public async Task AddGroupAsync_ForLocalConnection_AlreadyInGroup_SkipsDuplicate()
         {
-            var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
+            var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
 
             using (var client = new TestClient())
             {
@@ -77,7 +75,7 @@ namespace SignalR.Orleans.Tests
                 await manager.AddToGroupAsync(connection.ConnectionId, "dmx");
                 await manager.AddToGroupAsync(connection.ConnectionId, "dmx");
 
-                var grain = _fixture.ClientProvider.GetClient().GetGroupGrain("MyHub", "dmx");
+                var grain = _fixture.Client.GetGroupGrain("MyHub", "dmx");
                 var result = await grain.Count();
                 Assert.Equal(1, result);
             }
@@ -86,21 +84,30 @@ namespace SignalR.Orleans.Tests
         [Fact]
         public async Task DisconnectConnection_RemovesConnection_FromGroup()
         {
-            var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
+            var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
 
             using (var client = new TestClient())
             {
+                var groupGrain = _fixture.Client.GetGroupGrain("MyHub", "dre");
+
                 var connection = HubConnectionContextUtils.Create(client.Connection);
-
                 await manager.OnConnectedAsync(connection);
-
                 await manager.AddToGroupAsync(connection.ConnectionId, "dre");
+
+                Assert.Equal(1, await groupGrain.Count());
 
                 await manager.OnDisconnectedAsync(connection);
 
-                var grain = _fixture.ClientProvider.GetClient().GetGroupGrain("MyHub", "dre");
-                var result = await grain.Count();
-                Assert.Equal(0, result);
+                // Give the disconnection stream message time to be processed.
+                var sw = Stopwatch.StartNew();
+                while (await groupGrain.Count() != 0)
+                {
+                    if (sw.ElapsedMilliseconds > 1000) throw new Exception("User was not cleared from group after disconnection.");
+                    await Task.Delay(20);
+                }
+
+                //var result = await groupGrain.Count();
+                //Assert.Equal(0, result);
             }
         }
 
@@ -109,7 +116,7 @@ namespace SignalR.Orleans.Tests
         {
             using (var client1 = new TestClient())
             {
-                var manager = new OrleansHubLifetimeManager<DaHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<DaHub>>(), _fixture.ClientProvider);
+                var manager = new OrleansHubLifetimeManager<DaHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<DaHub>>(), _fixture.Client);
 
                 var connection1 = HubConnectionContextUtils.Create(client1.Connection);
 
@@ -126,7 +133,7 @@ namespace SignalR.Orleans.Tests
         {
             using (var client1 = new TestClient())
             {
-                var manager = new OrleansHubLifetimeManager<DaHubx>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<DaHubx>>(), _fixture.ClientProvider);
+                var manager = new OrleansHubLifetimeManager<DaHubx>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<DaHubx>>(), _fixture.Client);
 
                 var connection1 = HubConnectionContextUtils.Create(client1.Connection);
 
@@ -143,7 +150,7 @@ namespace SignalR.Orleans.Tests
         {
             using (var client1 = new TestClient())
             {
-                var manager = new OrleansHubLifetimeManager<DaHubUsingBase>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<DaHubUsingBase>>(), _fixture.ClientProvider);
+                var manager = new OrleansHubLifetimeManager<DaHubUsingBase>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<DaHubUsingBase>>(), _fixture.Client);
 
                 var connection1 = HubConnectionContextUtils.Create(client1.Connection);
 
@@ -161,7 +168,7 @@ namespace SignalR.Orleans.Tests
             using (var client1 = new TestClient())
             using (var client2 = new TestClient())
             {
-                var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
+                var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
 
                 var connection1 = HubConnectionContextUtils.Create(client1.Connection);
                 var connection2 = HubConnectionContextUtils.Create(client2.Connection);
@@ -182,9 +189,9 @@ namespace SignalR.Orleans.Tests
         [Fact]
         public async Task InvokeAllAsync_ForSpecificHub_WithMultipleServers_WritesTo_AllConnections_Output()
         {
-            var manager1 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
-            var manager2 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
-            var manager3 = new OrleansHubLifetimeManager<DifferentHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<DifferentHub>>(), _fixture.ClientProvider);
+            var manager1 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
+            var manager2 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
+            var manager3 = new OrleansHubLifetimeManager<DifferentHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<DifferentHub>>(), _fixture.Client);
 
             using (var client1 = new TestClient())
             using (var client2 = new TestClient())
@@ -209,8 +216,8 @@ namespace SignalR.Orleans.Tests
         [Fact]
         public async Task InvokeAllAsync_WithMultipleServers_DoesNotWrite_ToDisconnectedConnections_Output()
         {
-            var manager1 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
-            var manager2 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
+            var manager1 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
+            var manager2 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
 
             using (var client1 = new TestClient())
             using (var client2 = new TestClient())
@@ -234,8 +241,8 @@ namespace SignalR.Orleans.Tests
         [Fact]
         public async Task InvokeAllAsync_WithMultipleServers_WritesToAllConnections_Output()
         {
-            var manager1 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
-            var manager2 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
+            var manager1 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
+            var manager2 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
 
             using (var client1 = new TestClient())
             using (var client2 = new TestClient())
@@ -259,7 +266,7 @@ namespace SignalR.Orleans.Tests
             using (var client1 = new TestClient())
             using (var client2 = new TestClient())
             {
-                var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
+                var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
 
                 var connection1 = HubConnectionContextUtils.Create(client1.Connection);
                 var connection2 = HubConnectionContextUtils.Create(client2.Connection);
@@ -280,7 +287,7 @@ namespace SignalR.Orleans.Tests
             using (var client1 = new TestClient())
             using (var client2 = new TestClient())
             {
-                var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
+                var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
                 var connection1 = HubConnectionContextUtils.Create(client1.Connection);
                 var connection2 = HubConnectionContextUtils.Create(client2.Connection);
 
@@ -292,7 +299,7 @@ namespace SignalR.Orleans.Tests
 
                 await manager.SendGroupAsync(groupName, "Hello", new object[] { "World" });
 
-                var grain = _fixture.ClientProvider.GetClient().GetGroupGrain("MyHub", groupName);
+                var grain = _fixture.Client.GetGroupGrain("MyHub", groupName);
                 var connectionsCount = await grain.Count();
 
                 await AssertMessageAsync(client1);
@@ -300,6 +307,9 @@ namespace SignalR.Orleans.Tests
 
                 await manager.SendGroupAsync(groupName, "Hello", new object[] { "World" });
                 await manager.SendGroupAsync(groupName, "Hello", new object[] { "World" });
+
+                // Give the disconnection stream message time to be processed by the group grain.
+                await Task.Delay(1000);
 
                 connectionsCount = await grain.Count();
                 Assert.Equal(1, connectionsCount);
@@ -309,8 +319,8 @@ namespace SignalR.Orleans.Tests
         [Fact]
         public async Task InvokeConnectionAsync_ForLocalConnection_DoesNotPublish()
         {
-            var manager1 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
-            var manager2 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
+            var manager1 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
+            var manager2 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
 
             using (var client = new TestClient())
             {
@@ -331,17 +341,17 @@ namespace SignalR.Orleans.Tests
         public async Task InvokeConnectionAsync_OnNonExistentConnection_DoesNotThrow()
         {
             var invalidConnection = "NotARealConnectionId";
-            var grain = _fixture.ClientProvider.GetClient().GetClientGrain("MyHub", invalidConnection);
+            var grain = _fixture.Client.GetClientGrain("MyHub", invalidConnection);
             await grain.OnConnect(Guid.NewGuid());
-            var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
+            var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
             await manager.SendConnectionAsync(invalidConnection, "Hello", new object[] { "World" });
         }
 
         [Fact]
         public async Task InvokeConnectionAsync_OnServer_WithoutConnection_WritesOutputTo_Connection()
         {
-            var manager1 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
-            var manager2 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
+            var manager1 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
+            var manager2 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
 
             using (var client = new TestClient())
             {
@@ -360,7 +370,7 @@ namespace SignalR.Orleans.Tests
         {
             using (var client = new TestClient())
             {
-                var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
+                var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
                 var connection = HubConnectionContextUtils.Create(client.Connection);
 
                 await manager.OnConnectedAsync(connection);
@@ -377,7 +387,7 @@ namespace SignalR.Orleans.Tests
             using (var client1 = new TestClient())
             using (var client2 = new TestClient())
             {
-                var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
+                var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
                 var connection1 = HubConnectionContextUtils.Create(client1.Connection);
                 var connection2 = HubConnectionContextUtils.Create(client2.Connection);
 
@@ -394,8 +404,8 @@ namespace SignalR.Orleans.Tests
         [Fact]
         public async Task InvokeGroupAsync_OnServer_WithoutConnection_WritesOutputTo_GroupConnection()
         {
-            var manager1 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
-            var manager2 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
+            var manager1 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
+            var manager2 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
 
             using (var client = new TestClient())
             {
@@ -418,7 +428,7 @@ namespace SignalR.Orleans.Tests
             using (var client2 = new TestClient())
             using (var client3 = new TestClient())
             {
-                var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
+                var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
                 var connection1 = HubConnectionContextUtils.Create(client1.Connection);
                 var connection2 = HubConnectionContextUtils.Create(client2.Connection);
                 var connection3 = HubConnectionContextUtils.Create(client3.Connection);
@@ -444,7 +454,7 @@ namespace SignalR.Orleans.Tests
             using (var client1 = new TestClient())
             using (var client2 = new TestClient())
             {
-                var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
+                var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
                 var connection1 = HubConnectionContextUtils.Create(client1.Connection);
                 var connection2 = HubConnectionContextUtils.Create(client2.Connection);
 
@@ -468,7 +478,7 @@ namespace SignalR.Orleans.Tests
             using (var client1 = new TestClient())
             using (var client2 = new TestClient())
             {
-                var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
+                var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
                 var connection1 = HubConnectionContextUtils.Create(client1.Connection);
                 var connection2 = HubConnectionContextUtils.Create(client2.Connection);
 
@@ -491,7 +501,7 @@ namespace SignalR.Orleans.Tests
             using (var client1 = new TestClient())
             using (var client2 = new TestClient())
             {
-                var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
+                var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
                 var connection1 = HubConnectionContextUtils.Create(client1.Connection);
                 var connection2 = HubConnectionContextUtils.Create(client2.Connection);
 
@@ -511,8 +521,8 @@ namespace SignalR.Orleans.Tests
         [Fact]
         public async Task RemoveGroup_FromConnection_OnDifferentServer_NotInGroup_DoesNothing()
         {
-            var manager1 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
-            var manager2 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
+            var manager1 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
+            var manager2 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
 
             using (var client = new TestClient())
             {
@@ -527,7 +537,7 @@ namespace SignalR.Orleans.Tests
         [Fact]
         public async Task RemoveGroup_FromLocalConnection_NotInGroup_DoesNothing()
         {
-            var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
+            var manager = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
 
             using (var client = new TestClient())
             {
@@ -542,8 +552,8 @@ namespace SignalR.Orleans.Tests
         [Fact]
         public async Task RemoveGroupAsync_ForConnection_OnDifferentServer_Works()
         {
-            var manager1 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
-            var manager2 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.ClientProvider);
+            var manager1 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
+            var manager2 = new OrleansHubLifetimeManager<MyHub>(new LoggerFactory().CreateLogger<OrleansHubLifetimeManager<MyHub>>(), _fixture.Client);
 
             using (var client = new TestClient())
             {
@@ -570,7 +580,7 @@ namespace SignalR.Orleans.Tests
             var message = Assert.IsType<InvocationMessage>(await client.ReadAsync().OrTimeout());
             Assert.Equal("Hello", message.Target);
             Assert.Single(message.Arguments);
-            Assert.Equal("World", message.Arguments[0].ToString());
+            Assert.Equal("World", message.Arguments[0]?.ToString());
         }
     }
 }
